fix(api): hallazgos de code-review previos al PR a develop
Bloqueantes de CI: ruff format en 6 ficheros y un float sin castear en
accent_palette.py que rompía mypy.

Hallazgos reales:
- El PUT de /admin/analytics-settings podía devolver 200 sin persistir
  nada si la fila 'default' no existía (mutar y hacer flush() sobre un
  objeto transitorio no escribe en BD) — se añade session.add().
- Índice por created_at en cookie_consents: el agregado semanal de
  consentimientos filtraba por rango de fecha sin ningún índice de
  apoyo (solo existía el de organization_id, que esa consulta ni usa).

+++ apps/api/app/modules/theme_templates/accent_palette.py
@@ -71,7 +71,7 @@ def hex_a_oklch(hex_color: str) -> tuple[float, float, float]:
         proporcion = canal_0_255 / 255
         if proporcion <= 0.04045:
             return proporcion / 12.92
-        return ((proporcion + 0.055) / 1.055) ** 2.4
+        return float(((proporcion + 0.055) / 1.055) ** 2.4)
 
     r, g, b = (lineal(canal) for canal in rgb)
